+++ src/main/java/com/appdynamics/extensions/redis/RedisMonitor.java
@@ -1,28 +1,8 @@
-/**
- * Copyright 2017 AppDynamics, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/*package com.appdynamics.extensions.redis;
+package com.appdynamics.extensions.redis;
 
-import com.appdynamics.extensions.conf.MonitorConfiguration;
+import com.appdynamics.extensions.ABaseMonitor;
+import com.appdynamics.extensions.ATaskExecutor;
 import com.appdynamics.extensions.util.AssertUtils;
-import com.appdynamics.extensions.util.MetricWriteHelper;
-import com.appdynamics.extensions.util.MetricWriteHelperFactory;
-import com.singularity.ee.agent.systemagent.api.AManagedMonitor;
-import com.singularity.ee.agent.systemagent.api.TaskExecutionContext;
-import com.singularity.ee.agent.systemagent.api.TaskOutput;
-import com.singularity.ee.agent.systemagent.api.exception.TaskExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -31,50 +11,36 @@ import java.util.Map;
 
 import static com.appdynamics.extensions.redis.utils.Constants.DEFAULT_METRIC_PREFIX;
 
-public class  RedisMonitor extends AManagedMonitor {
+public class RedisMonitor extends ABaseMonitor {
+
     private static final Logger logger = LoggerFactory.getLogger(RedisMonitor.class);
-    private MonitorConfiguration configuration;
     private static long previousTimeStamp = System.currentTimeMillis();
     private static long currentTimeStamp = System.currentTimeMillis();
 
-    public RedisMonitor(){
-        logger.info("Using Redis Monitor Version [" + getImplementationVersion() + "]");
+    @Override
+    protected String getDefaultMetricPrefix() {
+        return DEFAULT_METRIC_PREFIX;
     }
 
-    public TaskOutput execute(Map<String, String> var1, TaskExecutionContext var2) throws TaskExecutionException{
-        logger.debug("The raw arguments are {}" + var1);
-        logger.debug("The previousTimeStamp : {} and the currentTimeStamp : {}",previousTimeStamp, currentTimeStamp);
-        initialize(var1);
-        configuration.executeTask();
-        configuration.getMetricWriter().onTaskComplete();
-        previousTimeStamp = currentTimeStamp;
-        currentTimeStamp = System.currentTimeMillis();
-        return new TaskOutput("Redis monitor run completed successfully.");
+    @Override
+    protected String getMonitorName() {
+        return "Redis Monitor";
     }
 
-    private void initialize(Map<String, String> var1) {
-        if(configuration == null){
-            MetricWriteHelper metricWriteHelper = MetricWriteHelperFactory.create(this);
-            MonitorConfiguration conf = new MonitorConfiguration(DEFAULT_METRIC_PREFIX, new TaskRunner(), metricWriteHelper);
-            conf.setConfigYml(var1.get("config-file"));
-            this.configuration = conf;
+    @Override
+    protected void doRun(ATaskExecutor taskExecutor) {
+        List<Map<String,String>> servers = (List<Map<String,String>>)configuration.getConfigYml().get("servers");
+        AssertUtils.assertNotNull(servers, "The 'servers' section in config.yml is not initialised");
+        for (Map<String, String> server : servers) {
+            RedisMonitorTask task = new RedisMonitorTask(configuration, server, previousTimeStamp, currentTimeStamp);
+            taskExecutor.submit(server.get("name"),task);
         }
     }
 
-    private class TaskRunner implements Runnable {
-        public void run(){
-            List<Map<String,String>> servers = (List<Map<String,String>>)configuration.getConfigYml().get("servers");
-            AssertUtils.assertNotNull(servers, "The 'servers' section in config.yml is not initialised");
-            for (Map<String, String> server : servers) {
-                RedisMonitorTask task = new RedisMonitorTask(configuration, server, previousTimeStamp, currentTimeStamp);
-                configuration.getExecutorService().execute(task);
-            }
-        }
+    @Override
+    protected int getTaskCount() {
+        List<Map<String,String>> servers = (List<Map<String,String>>)configuration.getConfigYml().get("servers");
+        AssertUtils.assertNotNull(servers, "The 'servers' section in config.yml is not initialised");
+        return servers.size();
     }
-
-    private static String getImplementationVersion() {
-        return RedisMonitor.class.getPackage().getImplementationTitle();
-    }
-
 }
-*/
