+++ src/main/java/com/appdynamics/extensions/redis/metrics/SlowLogMetrics.java
@@ -1,8 +1,8 @@
 package com.appdynamics.extensions.redis.metrics;
 
 import com.appdynamics.extensions.conf.MonitorConfiguration;
+import com.appdynamics.extensions.metrics.Metric;
 import com.appdynamics.extensions.util.AssertUtils;
-import com.appdynamics.extensions.util.Metric;
 import com.google.common.collect.Lists;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,11 +13,10 @@ import java.util.List;
 import java.util.ListIterator;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
-
 import static com.appdynamics.extensions.redis.utils.Constants.METRIC_SEPARATOR;
-import static com.appdynamics.extensions.redis.utils.Constants.transformAndPrintNodeLevelMetrics;
 
 public class SlowLogMetrics implements Runnable {
+
     private JedisPool jedisPool;
     private Map<String, ?> metricsMap;
     private MonitorConfiguration configuration;
@@ -47,7 +46,7 @@ public class SlowLogMetrics implements Runnable {
         extractSlowLogPropertiesMap(slowLogMetricsList);
         finalMetricList = extractSlowLogMetricsList();
         logger.debug("Printing SlowLog metrics for server {}", server.get("name"));
-        transformAndPrintNodeLevelMetrics(configuration, finalMetricList);
+        configuration.getMetricWriter().transformAndPrintNodeLevelMetrics(finalMetricList);
         countDownLatch.countDown();
     }
