+++ src/main/java/com/appdynamics/extensions/redis/utils/InfoMapExtractor.java
@@ -2,7 +2,6 @@ package com.appdynamics.extensions.redis.utils;
 
 import com.google.common.base.Splitter;
 import com.google.common.collect.Maps;
-
 import java.util.Map;
 
 public class InfoMapExtractor {
