+++ src/test/java/com/appdynamics/extensions/redis/metrics/SlowLogMetricsTest.java
@@ -1,9 +1,9 @@
 package com.appdynamics.extensions.redis.metrics;
 
+import com.appdynamics.extensions.MetricWriteHelper;
+import com.appdynamics.extensions.MetricWriteHelperFactory;
 import com.appdynamics.extensions.conf.MonitorConfiguration;
-import com.appdynamics.extensions.util.Metric;
-import com.appdynamics.extensions.util.MetricWriteHelper;
-import com.appdynamics.extensions.util.MetricWriteHelperFactory;
+import com.appdynamics.extensions.metrics.Metric;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.singularity.ee.agent.systemagent.api.AManagedMonitor;
@@ -19,16 +19,15 @@ import org.slf4j.LoggerFactory;
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
 import redis.clients.util.Slowlog;
-
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CountDownLatch;
-
 import static org.mockito.Mockito.*;
 
 public class SlowLogMetricsTest {
+
     private class TaskRunner implements Runnable{
         public void run(){
 
@@ -61,7 +60,6 @@ public class SlowLogMetricsTest {
         list11.add("slowlog".getBytes());
         list11.add("get".getBytes());
         list1.add(list11);
-
         List list2 = new ArrayList();
         list2.add(14L);
         list2.add(1505158203L);
@@ -70,10 +68,8 @@ public class SlowLogMetricsTest {
         list21.add("slowlog".getBytes());
         list21.add("get".getBytes());
         list2.add(list21);
-
         objectList.add(list1);
         objectList.add(list2);
-
         when(jedis.slowlogGet(jedis.slowlogLen())).thenReturn(Slowlog.from(objectList));
         Map<String, String> server = Maps.newHashMap();
         server.put("host", "localhost");
@@ -107,7 +103,6 @@ public class SlowLogMetricsTest {
         list11.add("slowlog".getBytes());
         list11.add("get".getBytes());
         list1.add(list11);
-
         List list2 = new ArrayList();
         list2.add(14L);
         list2.add(1505158203L);
@@ -116,7 +111,6 @@ public class SlowLogMetricsTest {
         list21.add("slowlog".getBytes());
         list21.add("get".getBytes());
         list2.add(list21);
-
         List list3 = new ArrayList();
         list3.add(13L);
         list3.add(1505158201L);
@@ -125,11 +119,9 @@ public class SlowLogMetricsTest {
         list31.add("slowlog".getBytes());
         list31.add("get".getBytes());
         list3.add(list31);
-
         objectList.add(list1);
         objectList.add(list2);
         objectList.add(list3);
-
         when(jedis.slowlogGet(jedis.slowlogLen())).thenReturn(Slowlog.from(objectList));
         Map<String, String> server = Maps.newHashMap();
         server.put("host", "localhost");
@@ -137,8 +129,6 @@ public class SlowLogMetricsTest {
         server.put("name", "Server1");
         SlowLogMetrics slowLogMetrics = new SlowLogMetrics(monitorConfiguration, server, jedisPool, countDownLatch, 1505158200000L, 1505158220000L);
         slowLogMetrics.run();
-
-
         List<Object> objectList2 = Lists.newArrayList();
         List list4 = new ArrayList();
         list4.add(15L);
@@ -148,7 +138,6 @@ public class SlowLogMetricsTest {
         list41.add("slowlog".getBytes());
         list41.add("get".getBytes());
         list4.add(list41);
-
         List list5 = new ArrayList();
         list5.add(14L);
         list5.add(1505158203L);
@@ -157,11 +146,8 @@ public class SlowLogMetricsTest {
         list51.add("slowlog".getBytes());
         list51.add("get".getBytes());
         list5.add(list51);
-
-
         objectList2.add(list4);
         objectList2.add(list5);
-
         when(jedis.slowlogGet(jedis.slowlogLen())).thenReturn(Slowlog.from(objectList2));
         Map<String, String> server2 = Maps.newHashMap();
         server2.put("host", "localhost");
@@ -169,11 +155,9 @@ public class SlowLogMetricsTest {
         server2.put("name", "Server2");
         SlowLogMetrics slowLogMetrics2 = new SlowLogMetrics(monitorConfiguration, server2, jedisPool, countDownLatch, 1505158200000L, 1505158220000L);
         slowLogMetrics2.run();
-
         metricWriteHelper.onTaskComplete();
         verify(metricWriteHelper,times(3)).printMetric(pathCaptor.capture());
         int count = 0;
-
         for (List<Metric> metricList : pathCaptor.getAllValues()){
             for(Metric metric : metricList){
                 if(metric.getMetricPath().equals("Server|Component:AppLevels|Custom Metrics|Redis|Cluster|SlowLog|no_of_new_slow_logs")){
