+++ src/test/java/com/appdynamics/extensions/redis/utils/InfoMapExtractorTest.java
@@ -4,15 +4,14 @@ import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-
 import java.io.File;
 import java.io.IOException;
 import java.util.Map;
-
 /**
  * Created by venkata.konala on 8/4/17.
  */
 public class InfoMapExtractorTest {
+
     InfoMapExtractor infoMapExtractor = new InfoMapExtractor();
     String info;
     @Before
@@ -35,6 +34,4 @@ public class InfoMapExtractorTest {
         Map<String, String> sectionInfoMap3 = infoMapExtractor.extractInfoAsHashMap(info, "No");
         Assert.assertTrue(sectionInfoMap3.size() == 0);
     }
-
-
 }
